--- docs/src/stories/ui-patterns/ActionList/TreeViewReadMe.stories.mdx
@@ -1,30 +1,0 @@
-import {Meta} from '@storybook/addon-docs'
-
-<Meta title="UI Patterns/ActionList/ActionTreeView/TreeViewListItemCollapsible/README" />
-
-# Action list tree view variant
-
-`TreeView` markup and styles is considered unstable and in active development. The `collapsible` functionality and markup differs slightly from `ActionListItemCollapsible`, though this will likely converge as we work towards stability.
-
-**All breaking changes have been scoped to `.ActionList` and reverted in `.ActionList--tree` for backwards compatibility**
-
-### ActionListItemCollapsible
-
-| element    | attributes                                                |
-| ---------- | --------------------------------------------------------- |
-| `<li>`     | classes: `ActionList-item`, `ActionList-item--hasSubItem` |
-| `<button>` | `onClick`, `aria-expanded`, `aria-controls`. `id`         |
-| `<button>` | classes: `ActionList-content--hasActiveSubItem`           |
-
-### TreeViewListItemCollapsible
-
-| element | attributes                                                                                                                 |
-| ------- | -------------------------------------------------------------------------------------------------------------------------- |
-| `<li>`  | `onClick`, `aria-level`, `role="treeitem"`, `aria-expanded`, `id`                                                          |
-| `<li>`  | classes: `ActionList-item`, `ActionList-item--hasSubItem`, `ActionList-item--subItem`, `ActionList-item--hasActiveSubItem` |
-
-For `ActionListItemCollapsible` click handling functionality moved down one level to a semantic `<button>` element. With that change, CSS expand/collapse logic also moved down a level. This is indicated with our BEM style naming convention `item--` vs `content--`
-
-| original class                      | new class                              |
-| ----------------------------------- | -------------------------------------- |
-| `ActionList-item--hasActiveSubItem` | `ActionList-content--hasActiveSubItem` |
